compute fermat spiral

+++ ws/src/control/include/control/path/path_line.hpp
@@ -18,9 +18,7 @@ public:
 
     double course_error(const ignition::math::Vector2d& pos, double course) override;
 
-    std::vector<ignition::math::Vector2d> sample(size_t number_of_samples) override;
+    std::vector<ignition::math::Vector2d> sample(int number_of_samples) override;
 private:
     const ignition::math::Vector2d pos_begin, pos_end;
 };
-
-
